Refactor dependencies, remove duplicates, and clean up debug logs
