+++ backend/app/api/v1/endpoints/serial.py
@@ -3,39 +3,18 @@ Serial Communication API Endpoints for AT Commands
 """
 
 import logging
-from fastapi import APIRouter, Request, Header, Depends, HTTPException
+from fastapi import APIRouter, Request, Depends
 from typing import Optional
 
 from app.core.response import APIResponse
-from app.core.exceptions import SerialException, SessionException, ErrorCode
+from app.core.dependencies import validate_session_dependency
 from app.services.serial_service import serial_service
-from app.services.session_service import session_service
 from app.schemas.serial_schemas import (
     SerialConfig, RawDataRequest
 )
 
 logger = logging.getLogger(__name__)
 router = APIRouter()
-
-
-async def get_session_id_from_header(x_session_id: Optional[str] = Header(None)) -> Optional[str]:
-    """从请求头获取会话ID"""
-    return x_session_id
-
-
-async def validate_session_dependency(
-    request: Request,
-    session_id: Optional[str] = Depends(get_session_id_from_header)
-) -> str:
-    """会话验证依赖"""
-    if not session_id:
-        raise HTTPException(status_code=401, detail="缺少会话ID，请先登录")
-    
-    is_valid = await session_service.validate_session(session_id, request)
-    if not is_valid:
-        raise HTTPException(status_code=401, detail="会话无效或已过期，请重新登录")
-    
-    return session_id
 
 
 @router.get("/ports", response_model=APIResponse)
